+++ 1Week_Mission/src/main/java/com/ll/exam/mutbooks/app/security/handler/CustomAuthSuccessHandler.java
@@ -1,7 +1,7 @@
 package com.ll.exam.mutbooks.app.security.handler;
 
 import com.ll.exam.mutbooks.app.security.dto.MemberContext;
-import com.ll.exam.mutbooks.app.util.Ut;
+import com.ll.exam.mutbooks.util.Ut;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.web.DefaultRedirectStrategy;
 import org.springframework.security.web.RedirectStrategy;
